Clean up console logs, rename test file
